upm-c: remove uint8_t limit from pin/bus variables in structs
This removes unnecessary uint8 types from C driver structs that prevented some sensors from being initialized when connected to a MRAA subplatform.

+++ src/m24lr64e/m24lr64e.h
@@ -113,9 +113,9 @@ typedef enum {
  */
 typedef struct _m24lr64e_context {
     mraa_i2c_context            i2c;
-    uint8_t                     bus;
+    int                         bus;
     uint8_t                     address;
-    m24lr64e_access_mode    mode;
+    m24lr64e_access_mode        mode;
 } *m24lr64e_context;
 
 /**
